requirements.txt
# --- Core Web Framework ---
fastapi
uvicorn
gunicorn[uvloop]

# --- AI & LangChain (Downgraded to stable 0.3.7 to fix AgentExecutor error) ---
langchain==0.3.7
langchain-community==0.3.7
langchain-experimental
langchain-google-genai
langchain-neo4j
langchain-text-splitters
pypdf

# --- Databases ---
supabase
neo4j

# --- Utilities ---
python-dotenv
pydantic
tiktoken
python-dateutil
pytz
email_validator
twilio
apscheduler

# --- Google Calendar API ---
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

unstructured
markdown